refactor(room): move dao classes into own package
