+++ app/src/main/java/com/chesire/zwei/room/dao/MountDao.kt
@@ -1,4 +1,4 @@
-package com.chesire.zwei.room
+package com.chesire.zwei.room.dao
 
 import android.arch.persistence.room.Dao
 import android.arch.persistence.room.Insert
